requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
openpyxl>=3.1.0
reportlab>=4.0.0
openai>=1.3.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
